#534 don't send delete jobs request if no jobs are selected

+++ quick-start/src/main/ui/app/jobs/jobs.component.ts
@@ -191,13 +191,15 @@ export class JobsComponent implements OnChanges, OnDestroy, OnInit {
   }
 
   deleteJobs() {
-    this.jobService.deleteJobs(this.jobsToDelete)
-      .subscribe(response => {
-          this.getJobs();
-        },
-        () => {
-          this.dialogService.alert("Failed to delete jobs");
-        });
+    if (this.jobsToDelete.length > 0) {
+      this.jobService.deleteJobs(this.jobsToDelete)
+        .subscribe(response => {
+            this.getJobs();
+          },
+          () => {
+            this.dialogService.alert("Failed to delete jobs");
+          });
+    }
   }
 
   render(o) {
